requirements.txt
streamlit==1.24.0
numpy==1.23.5
opencv-python==4.8.0
joblib==1.3.1
Pillow==9.5.0
tensorflow==2.13.0
matplotlib==3.7.2
seaborn==0.12.2
